#2514 Change method of getting quiz name

+++ service/src/main/java/skills/services/UserAchievementsAndPointsManagement.groovy
@@ -256,9 +256,11 @@ class UserAchievementsAndPointsManagement {
         assert levels
         if (levels) {
             levels.each {
-                int numUpdated = userAchievedLevelRepo.removeProjectLevelAchievementsIfUsersDoNotQualify(projectId, it.level, it.pointsFrom)
-                log.info("Remove project's level achievements for projectId=[{}], level=[{}], pointsFromExclusive=[{}]. Num rows removed = [{}]",
-                        projectId, it.level, it.pointsFrom, numUpdated)
+                if(it.pointsFrom) {
+                    int numUpdated = userAchievedLevelRepo.removeProjectLevelAchievementsIfUsersDoNotQualify(projectId, it.level, it.pointsFrom)
+                    log.info("Remove project's level achievements for projectId=[{}], level=[{}], pointsFromExclusive=[{}]. Num rows removed = [{}]",
+                            projectId, it.level, it.pointsFrom, numUpdated)
+                }
             }
         }
     }
